refactor(clients): split _wire_models into per-app submodules
The flat 283-line _wire_models.py held 14 Pydantic models covering
every *arr family.  A future per-app drift (for example a Sonarr
v4 field rename) had no natural home; every change forced a walk
through the whole file.

Split the module into a package:

  clients/_wire_models/
    __init__.py     re-exports every public name (__all__ parity)
    common.py       _ArrModel, PaginatedResponse, SystemStatus,
                    QueueStatus, Arr{Series,Artist,Author}, and the
                    internal _Wire{Episode,Movie}File /
                    _Wire{Album,Book}Statistics helpers
    sonarr.py       SonarrWantedEpisode, SonarrLibraryEpisode
    radarr.py       RadarrWantedMovie, RadarrLibraryMovie
    lidarr.py       LidarrWantedAlbum, LidarrLibraryAlbum
    readarr.py      ReadarrWantedBook, ReadarrLibraryBook
    whisparr_v2.py  WhisparrV2WantedEpisode, WhisparrV2LibraryEpisode
    whisparr_v3.py  WhisparrV3LibraryMovie

Every ``Field(alias="camelCase")`` stays byte-equal; the package
root stays import-compatible with the pre-split flat module so
the 9 call sites (base.py + 7 per-app clients + several tests)
need no update.
